Reorganize modules: move BlockBindings, EnclosingSymbol; rename types
Move BlockBindings to values.rs (where Value lives, which it owns
values of), and EnclosingSymbol to env.rs (where Stack frames
reference it). Rename types.rs to type_defs.rs to better reflect its
contents (TypeDef and TypeDefAndMethods).
